Initial Vagrant testing support, using the Chef provider with Chef Zero.
The Vagrant box is the default 12.04 and 13.04 provided by Opscode's Bento
project. It uses the Vagrant plugins vagrant-berkshelf, vagrant-cachier,
vagrant-chef-zero and vagrant-omnibus. Unfortunately Chef Zero does not
understand .rb files yet, so we're using the recipes directly in the run
list. This also means that searching by roles is broken so this needs to be
fixed ASAP.

Change-Id: I5fe8b175f15c4a21c16de46bc05ec6a0c27cfdaa