add each role for each network main component
In typical openstack configuration, there may be 3 types nodes:
controller node, network node, and compute node. From the network
point, it's always supposed that: the Quantum Server is on the
controller node; and the L2 agent, DHCP agent and L3 agent are on
the network node; the L2 agent are on the compute node. We need the
each role for each OpenStack network main component. This will make
the configuration more flexible. So I add them here.

Change-Id: Ida4826c6dfe770b58790d75561bea42e9584ca89
Closes-Bug: #1240509