Remove Berksfile.lock
Berksfile.lock is not updated so points to old versions of each of the
stackforge cookbooks. Removing this file ensures that a user gets the
most recent stackforge cookbooks when installing with berkshelf.

Add Berksfile.lock to .gitignore so that another version isn't
committed.

Update Readme with note about Berksfile.lock

Change-Id: I8ea3223e98ac4dbab94e45f2323c6dcea2b54504