Update repo with new testing suite
The new testing suite is based upon Chef 12, Chef DK and
Chef Provisioning.  It supports all in one and multi node
environments.  It will be used as the base for future gate
integration testing.  There's is more work to as noted in
the TODO list at the botton of the readme.

Note: to get this thru the existing gate, had to keep the existing
Gemfile.  Added a Gemfile-Provisioning for new stuff and updated
readme with some notes on that.  When we make the switch to kilo
rake gates, we can decide what type of gates the repo needs. I think
it's ready for rubocop at least and maybe continue with something
like spiecweasel to look at the role and environment syntax.

Added draft bootstrap and test default to rubocop to rakefile to
allow for easy integration with new gate rake jobs.

Change-Id: I187df8b78c4cecfb783ab26a6b66b8e8dd6745bd