Add auto create_key task to rakefile
To eliminate one more step in the process of running our repo steps,
added create_key task and made the test task depend upon it.

The other manual step is to run the berk_vendor task, but I hesitate
to include that in the test tasks as most developers are always messing
around with the code and would not like it blown away all the time.

Change-Id: I57c4fd8118c289386552e0ed1087959883724ec4