New test_patch helper script
This tool allows a single patch to be tested against the chef repo.
It will to do job of cloning the repo, merging in the patch(s),
setting up the test and running it, basic queries and tempest.

Also enhanced the Berksfile to allow for easy local development.

Usage: chef exec ruby test_patch.rb help test

Change-Id: I32ba57c29f2f5caee0391ea791ffe10fe06caceb