Fix issue with stdout not appearing on console for converge
Depending upon how the chef-client is envoked, it does a
STDOUT.tty? check to determine if output should be sent to
the doc formatter as STDOUT.  In many cases this check returns
false.  To override this, I added the --force-formatter option
to the chef-client so we can always get the output to STDOUT.
This will help with allowing batch testing to capture
all output in log files for a testing suite run.

Change-Id: Ibde012f9c3aa281ff1477c9b7e1d1ffa415e97cd