Allow env var REPO_OS to easily switch between platforms
Use a env var instead of manually editing repo files to use
centos.   This also makes it easy for new tools like, test-patch
to switch between the platforms.

Next step will be to allow REPO_DEVICE=  to easily
setup the network interfaces in the multi node cases.

I think env vars are a simple way to handle this, but we could also
consider an user config file ~/openstack-chef-repo.cnf.  Thoughts?

Change-Id: I82e82b35b51e08ff503fdb10fa1ce041c4be4af2