CI TEST
This will allow a basic all in one environment test to be run against
the openstack cookbook repo.  Once in place, it can also then be used
by individual cookbook patches to run the same tests.  This is just a
starting point for our CI work, see the todo list below.

- setup for using new experiemental gate
- run aio neutron environment
- run tests
- check for idempotency

TODOs
- get basic networking (local or flat) working on one nic
- gather logging info for failures (infra publish gate change?)
- multinode environment

Change-Id: I23ebc6092fc563ebfa0a9882668a5f51937a3013