CI - Depends-On support - Repo
This patch will use the Depends-On support enabled in the Common
bootstrap. The repo Berksfile already has support for REPO_DEV to
find cookbooks in a local path.  This change will enable the
repo to also be able to use the Depends-On commit message flag to
pull in dependent cookbooks.

To use Depends-On, add the following to your commit message AFTER
the description and BEFORE the change id and bug or blue print ref.
"Depends-On: <id>"
The id is the change id of the gerrit patch this patch depends on.

Change-Id: I7f9bc45995073ae210656835108b4e6a5257b082