Fix interface selection to be automatic
Added a list of possible interfaces for which the
vagrant provisioner will choose the first one that
is available.  No more editing the test recipes.

Cleaned up the os selection to hide the
version specific stuff to be consistent with the
environment logic.

Fixed up some file modes.

Forced the chef client config directory as it can search
in bad places in some scenarios.

Change-Id: I38a7a86b58893b912af99fabf04ca9052f914346