Fix memcached_servers setting for integration test
When running with chef-local, the search that is used to determine the
list of memcached_servers fails, leaving this setting empty for
integration tests. As a workaround, set the corresponding variable
explicitly in the allinone environmnet.

Also fix the Rakefile in order to allow the integration test to be
run more than once without issues.

Change-Id: I526de86e674477075c8ea51b31473d7a376723d7