Add support for kitchen testing
This changeset adds support for testing openstack-chef-repo with kitchen
and tempest.

Out of 63 tests, 33 pass and 30 are skipped, same as with current CI
tests.

The recipe, run_tempest in openstack-integration-test, is called at the
end of the installation phase.

In addition, a tempest run can be executed manually with
"kitchen verify".

Depends-On: I98706b1fc2c65a2359ca9b117a9491c0569514bd
Change-Id: I90c4ab627e3834ba9fd5791c425320e9536f7826